scripts/requirements.txt
websockets
cryptography
pytest
pytest-pudb
pytest-cov
pytest-asyncio
